+++ js/show-only-selected-languages.js
@@ -1,7 +1,26 @@
-// A globally exported function. Manipulates the page to only show code samples
-// for selected languages, hiding all other code samples (and headings).
-function showOnlySelectedLanguages(selectedLanguages) {
-  // RosettaCode pages are layed out roughly as follows:
+;(function () {
+  // This variable is closed over by multiple functions below.
+  var showingElements;
+
+  // A globally exported function. Manipulates the page to only show code samples
+  // for selected languages, hiding all other code samples (and headings).
+  window.showOnlySelectedLanguages = function showOnlySelectedLanguages(selectedLanguages) {
+    showOrHideEachTableOfContentsLink(selectedLanguages);
+    showOrHideEachContentSection(selectedLanguages);
+  }
+
+  function showOrHideEachTableOfContentsLink(selectedLanguages) {
+    var $tocLinks = $('#toc').find('.toclevel-1');
+
+    var test = function ($el) {
+      var language = $el.find('.toctext').text();
+      return isLanguageSelected(selectedLanguages, language);
+    }
+
+    showOrHideElements($tocLinks, test);
+  }
+
+  // The main content of RosettaCode pages are layed out roughly as follows:
   // ...
   // <h2>PHP</h2>
   // <pre>A PHP code sample</pre>
@@ -11,8 +30,8 @@ function showOnlySelectedLanguages(selectedLanguages) {
   // <pre>Another Python code sample</pre>
   // ...
   //
-  // The extension operates by iterating through DOM elements and alternating
-  // its `mode` between 'show' and 'hide', that is, whether to show or hide the
+  // The extension operates by iterating through DOM elements and switching a
+  // boolean `showingElements` flag, indicating whether to show or hide the
   // current and subsequent DOM elements. When the extension reaches an `h2`
   // tag, it checks whether that language is included in the set of languages
   // that the user wants displayed. If so, the `mode` is switched to 'show', and
@@ -20,41 +39,58 @@ function showOnlySelectedLanguages(selectedLanguages) {
   // If the language is not to be displayed, then the mode is set to 'hide',
   // until the next <h2> is reached, the `mode` is then set for that language,
   // etc.
-  var mode = 'hide';
+  function showOrHideEachContentSection(selectedLanguages) {
+    // the sibling elements of the table of contents (TOC) that come after the
+    // TOC are the main content sections (headings, code samples, etc.)
+    var $contentSections = $('#toc').find('~')
 
-  // iterate over sibling elements that come after the table of contents
-  $('#toc').find('~').each(function (_index, el) {
-    var $el = $(el);
+    var test = function ($el) {
+      // If this is an H2 element, then we are beginning a new section for a new
+      // language. We will want to update the `showingElements` flag for this
+      // and subsequent elements (until we reach the next H2, and change the
+      // mode again, etc.)
+      if ($el[0].tagName === 'H2') {
+        var language = $el.find('span.mw-headline').text();
+        if ( isLanguageSelected(selectedLanguages, language) ) {
+          return true;
+        } else {
+          return false
+        }
+      }
+      // continue in whatever mode we were in before
+      else {
+        return showingElements;
+      }
+    };
 
-    // If this is an H2 element, then we are beginning a new section for a new
-    // language. We will want to update the current mode (hide or show) for this
-    // and subsequent elements (until we reach the next H2, and change the mode
-    // again, etc.)
-    if (el.tagName === 'H2') {
-      setMode($el);
-    }
+    showOrHideElements($contentSections, test);
+  }
 
-    if (mode === 'hide') {
-      $el.addClass('rcls-hidden');
-    }
-    else {
-      $el.removeClass('rcls-hidden');
-    }
-  });
-
-  function setMode($h2) {
-    var language = $h2.find('span.mw-headline').text();
-    var showingLanguage = selectedLanguages.some(function (selectedLanguage) {
-      // Sometimes the languages are capitalized inconsistently
-      return language.toLowerCase() === selectedLanguage.toLowerCase();
+  // shouldShowTest is a function that will be passed a jQuerified version of
+  // each HTML element in the jQuery object. The function should return true or
+  // false for each element in turn to indicate whether the element should be
+  // shown.
+  function showOrHideElements($jqObject, shouldShowTest) {
+    $jqObject.each(function (_index, el) {
+      var $el = $(el);
+      showingElements = shouldShowTest($el);
+      showingElements ? $el.show() : $el.hide();
     });
-    mode = showingLanguage ? 'show' : 'hide';
+  }
 
+  function isLanguageSelected(selectedLanguages, language) {
     // special case; Mathematica and Wolfram share a single code sample
     if ( language === 'Mathematica / Wolfram Language' &&
-          ( selectedLanguages.indexOf('Mathematica') !== -1 ||
-             selectedLanguages.indexOf('Wolfram Language' ) !== -1) ) {
-      mode = 'show';
+          ( isLanguageSelected(selectedLanguages, 'Mathematica') ||
+             isLanguageSelected(selectedLanguages, 'Wolfram Language') ) ) {
+      return true;
     }
+
+    // normal case
+    return selectedLanguages.some(function (selectedLanguage) {
+      // sometimes the languages are capitalized inconsistently
+      return language.toLowerCase() === selectedLanguage.toLowerCase();
+    });
   }
-}
+
+}());